fix: Connect to bootstrap peers on startup (#2)
## Description

This PR fixes a devex issue where bootstrap instance would not be
connected to until a message was sent over the pubsub network. This
resulted it the bootstrap instance not knowing about the new instance
until later than expected.

The workaround for this was to call Connect instead of (or in addition
to) defining a bootstrap instance.

With this fix, the bootstrapped instance will know about the new
instance as soon as it goes online.